WIP Extract Operation (or ComputationAction?) from OperationMessage
