+++ syft/generic/pointers/object_pointer.py
@@ -12,7 +12,7 @@ from syft.generic.frameworks.hook.hook_args import register_backward_func
 from syft.generic.frameworks.hook import hook_args
 from syft.generic.frameworks.types import FrameworkTensor
 from syft.generic.object import AbstractObject
-from syft.messaging.message import ForceObjectDeleteMessage
+from syft.messaging.force_object_delete_message import ForceObjectDeleteMessage
 from syft.workers.abstract import AbstractWorker
 
 from syft.exceptions import RemoteObjectFoundError
