+++ syft/messaging/is_none_message.py
@@ -1,0 +1,38 @@
+import syft as sy
+
+from syft.messaging.message import Message
+from syft.workers.abstract import AbstractWorker
+
+
+class IsNoneMessage(Message):
+    """Check if a worker does not have an object with a specific id.
+
+    Occasionally we need to verify whether or not a remote worker has a specific
+    object. To do so, we send an IsNoneMessage, which returns True if the object
+    (such as a tensor) does NOT exist."""
+
+    # TODO: add more efficient detailer and simplifier custom for this type
+    # https://github.com/OpenMined/PySyft/issues/2512
+
+    def __init__(self, contents):
+        """Initialize the message using default Message constructor.
+
+        See Message.__init__ for details."""
+        super().__init__(contents)
+
+    @staticmethod
+    def detail(worker: AbstractWorker, msg_tuple: tuple) -> "IsNoneMessage":
+        """
+        This function takes the simplified tuple version of this message and converts
+        it into an IsNoneMessage. The simplify() method runs the inverse of this method.
+
+        Args:
+            worker (AbstractWorker): a reference to the worker necessary for detailing. Read
+                syft/serde/serde.py for more information on why this is necessary.
+            msg_tuple (Tuple): the raw information being detailed.
+        Returns:
+            ptr (IsNoneMessage): a IsNoneMessage.
+        Examples:
+            message = detail(sy.local_worker, msg_tuple)
+        """
+        return IsNoneMessage(sy.serde.msgpack.serde._detail(worker, msg_tuple[0]))
