+++ syft/messaging/object_request_message.py
@@ -1,0 +1,38 @@
+import syft as sy
+
+from syft.messaging.message import Message
+from syft.workers.abstract import AbstractWorker
+
+
+class ObjectRequestMessage(Message):
+    """Request another worker to send one of its objects
+
+    If ObjectMessage pushes an object to another worker, this Message type pulls an
+    object from another worker. It also assumes that the other worker will delete it's
+    local copy of the object after sending it to you."""
+
+    # TODO: add more efficient detailer and simplifier custom for this type
+    # https://github.com/OpenMined/PySyft/issues/2512
+
+    def __init__(self, contents):
+        """Initialize the message using default Message constructor.
+
+        See Message.__init__ for details."""
+        super().__init__(contents)
+
+    @staticmethod
+    def detail(worker: AbstractWorker, msg_tuple: tuple) -> "ObjectRequestMessage":
+        """
+        This function takes the simplified tuple version of this message and converts
+        it into an ObjectRequestMessage. The simplify() method runs the inverse of this method.
+
+        Args:
+            worker (AbstractWorker): a reference to the worker necessary for detailing. Read
+                syft/serde/serde.py for more information on why this is necessary.
+            msg_tuple (Tuple): the raw information being detailed.
+        Returns:
+            ptr (ObjectRequestMessage): a ObjectRequestMessage.
+        Examples:
+            message = detail(sy.local_worker, msg_tuple)
+        """
+        return ObjectRequestMessage(sy.serde.msgpack.serde._detail(worker, msg_tuple[0]))
