+++ syft/messaging/plan_command_message.py
@@ -1,0 +1,67 @@
+import syft as sy
+
+from syft.messaging.message import Message
+from syft.workers.abstract import AbstractWorker
+
+
+class PlanCommandMessage(Message):
+    """Message used to execute a command related to plans."""
+
+    # TODO: add more efficient detailer and simplifier custom for this type
+    # https://github.com/OpenMined/PySyft/issues/2512
+
+    def __init__(self, command_name: str, message: tuple):
+        """Initialize a PlanCommandMessage.
+
+        Args:
+            command_name (str): name used to identify the command.
+            message (Tuple): this is typically the args and kwargs of a method call on the client, but it
+                can be any information necessary to execute the command properly.
+        """
+
+        # call the parent constructor - setting the type integer correctly
+        super().__init__()
+
+        self.command_name = command_name
+        self.message = message
+
+    @property
+    def contents(self):
+        """Returns a tuple with the contents of the operation (backwards compatability)."""
+        return (self.command_name, self.message)
+
+    @staticmethod
+    def simplify(worker: AbstractWorker, ptr: "PlanCommandMessage") -> tuple:
+        """
+        This function takes the attributes of a PlanCommandMessage and saves them in a tuple
+
+        Args:
+            worker (AbstractWorker): a reference to the worker doing the serialization
+            ptr (PlanCommandMessage): a Message
+
+        Returns:
+            tuple: a tuple holding the unique attributes of the message
+        """
+        return (
+            sy.serde.msgpack.serde._simplify(worker, ptr.command_name),
+            sy.serde.msgpack.serde._simplify(worker, ptr.message),
+        )
+
+    @staticmethod
+    def detail(worker: AbstractWorker, msg_tuple: tuple) -> "PlanCommandMessage":
+        """
+        This function takes the simplified tuple version of this message and converts
+        it into a PlanCommandMessage. The simplify() method runs the inverse of this method.
+
+        Args:
+            worker (AbstractWorker): a reference to the worker necessary for detailing. Read
+                syft/serde/serde.py for more information on why this is necessary.
+            msg_tuple (Tuple): the raw information being detailed.
+        Returns:
+            ptr (PlanCommandMessage): a PlanCommandMessage.
+        """
+        command_name, message = msg_tuple
+        return PlanCommandMessage(
+            sy.serde.msgpack.serde._detail(worker, command_name),
+            sy.serde.msgpack.serde._detail(worker, message),
+        )
