+++ syft/serde/msgpack/serde.py
@@ -58,14 +58,14 @@ from syft.execution.plan import Plan
 from syft.execution.state import State
 from syft.execution.protocol import Protocol
 from syft.messaging.message import Message
-from syft.messaging.message import OperationMessage
-from syft.messaging.message import ObjectMessage
-from syft.messaging.message import ObjectRequestMessage
-from syft.messaging.message import IsNoneMessage
-from syft.messaging.message import GetShapeMessage
-from syft.messaging.message import ForceObjectDeleteMessage
-from syft.messaging.message import SearchMessage
-from syft.messaging.message import PlanCommandMessage
+from syft.messaging.operation_message import OperationMessage
+from syft.messaging.object_message import ObjectMessage
+from syft.messaging.object_request_message import ObjectRequestMessage
+from syft.messaging.is_none_message import IsNoneMessage
+from syft.messaging.get_shape_message import GetShapeMessage
+from syft.messaging.force_object_delete_message import ForceObjectDeleteMessage
+from syft.messaging.search_message import SearchMessage
+from syft.messaging.plan_command_message import PlanCommandMessage
 from syft.serde import compression
 from syft.serde.msgpack.native_serde import MAP_NATIVE_SIMPLIFIERS_AND_DETAILERS
 from syft.workers.abstract import AbstractWorker
